Added optional theming to Android screen
The previous method of allowing you to set the status bar color actually forced you to set a color, or the app would fail to build. This new method of theming overloads the `show` method to allow settting a custom theme, without breaking backwards compatability like the previous method had.